requirements.txt
asgiref==3.3.1
astroid==2.5.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-camel-case==1.2.0
dynaconf==3.1.4
isort==5.7.0
lazy-object-proxy==1.5.2
Markdown==3.3.4
mccabe==0.6.1
mimesis==4.1.3
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
